add record app move Answer model to record app
