+++ record/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from record.models import *
+
+# Register your models here.
+admin.site.register(Record)
+admin.site.register(Answer)
