+++ surveyer/models.py
@@ -15,16 +15,3 @@ class Survey(models.Model):
 
     def __unicode__(self):
         return Survey.name[:50]
-
-
-class Answer(models.Model):
-    mark = models.CharField(max_length=10, choices=(
-        ('G', ':-)'),
-        ('N', ':-|'),
-        ('B', ':-('),
-    ))
-    explanation = models.TextField(max_length=1000, null=True, blank=True)
-    question = models.ForeignKey(Question)
-
-    def __unicode__(self):
-        return Answer.mark + ' ' + Answer.question.__unicode__()
